arm64: dts: ti: k3-am65-mcu: Reserve some MCU SRAM for MCU R5F0
Add a child SRAM node to reserve a portion of the MCU domain
on-chip SRAM memory for use by the MCU R5F Core0. A preliminary
size of 256 KB (half of the current SRAM) is reserved to begin
with and can be adjusted as per need.

This memory is also added to the R5F0 remoteproc node using
the sram property.

Signed-off-by: Suman Anna <s-anna@ti.com>