arch: arm64: configs: Add a presilicon regular and a mini defconfig
Since there is development ongoing, use a different defconfig for
presilicon defconfigs. This allows the developers on integration
level to be able to work with configurations that depend on default
defconfig.

This is based off v4.14-rc1 with a bunch of driver cleanups and
platforms that have no meaning for TI presilicon development.

Signed-off-by: Nishanth Menon <nm@ti.com>