ti_config_fragments: rpmsg: Enable OMAP HwSpinlock driver
The OMAP HwSpinlock driver is reused for supporting the HwSpinlock
IP within K3 platforms. So, enable both the HwSpinlock core and the
OMAP HwSpinlock driver in the rpmsg config file. There are currently
no direct users of the HwSpinlock IP, but the built-in is the
preferred choice for this driver.

Signed-off-by: Suman Anna <s-anna@ti.com>