ti_config_fragments: rpmsg: Enable OMAP Mailbox driver
The OMAP Mailbox driver is reused for supporting the Mailbox
IP within K3 platforms. So, enable both the Mailbox core and the
OMAP Mailbox driver in the rpmsg config file. This is required
for remote processor messaging with various R5F remoteprocs on
the K3 AM65x and DRA82x family of SoCs. The Mailbox driver is
chosen to be built-in by default.

Signed-off-by: Suman Anna <s-anna@ti.com>