arm64: dts: ti: k3-am65: Add aliases for rproc nodes
Add aliases for both the R5F remoteproc processor cores present
within the MCU domain and common to all K3 AM65x boards. The aliases
uses the stem "rproc". Note that the R5F1 core and it's alias is
uniquely identified and used only if the MCU is running in Split
(non-LockStep) mode.

The aliases can be overridden, if needed, in the respective
derivative board dts files.

Signed-off-by: Suman Anna <s-anna@ti.com>