arm64: dts: ti: k3-dra8-main: Add mailbox cluster nodes
The DRA82x Main NavSS block contains a Mailbox IP instance with
multiple clusters. Each cluster is equivalent to an Mailbox IP
instance on OMAP platforms.

Add all the Mailbox clusters as their own nodes under the cbass_main
node instead of creating an almost empty parent node for the new
K3 mailbox IP and the clusters as its child nodes. All these nodes
are marked as disabled, and they need to be enabled along with the
appropriate child nodes on a need basis.

NOTE:
1. The NavSS only has a limited number of interrupts, so all the
   interrupts generated by a Mailbox IP are not added by default.
   Only the needed interrupts that are targeted towards the A72
   GIC will need to be be added later on when some sub-mailbox
   child nodes are added.
2. The nodes should be moved into a navss_main interconnect node if
   the local interconnect of Main NavSS is represented as a child
   node of the cbass_main interconnect node in the future.

Signed-off-by: Suman Anna <s-anna@ti.com>