arm64: dts: ti: dra822: Add IPC sub-mailbox nodes for all R5Fs & DSPs
Add the sub-mailbox nodes that are used to communicate between MPU and
various remote processors present in the DRA822 SoCs. These include the
R5F remote processors in the dual-R5F MCU cluster (MCU0) in the MCU domain
and the two dual-R5F MCU clusters (MCU1 & MCU2) in the MAIN domain; the
two C66x DSP remote processors and the single C7x DSP remote processor in
the MAIN domain. The parent mailbox cluster nodes are also enabled.

The sub-mailbox nodes utilize the System Mailbox clusters 0 through 4.
These sub-mailbox nodes are added to match the hard-coded mailbox
configuration used within the TI PDK IPC software package. The MCU
processor sub-systems are assumed to be running in Split mode, so a
sub-mailbox node is used by each of the R5F cores. The sub-mailbox node
for the first R5F core in each cluster is used in case of Lockstep mode.

NOTE:
The GIC_SPI interrupts to be used are usually dynamically allocated and
managed by the System Firmware through the ti-sci-irqchip driver. This
driver is currently not supported in 4.19 kernel, so the interrupt
properties in each of the enabled Mailbox clusters are not added. The
interrupt enablement for 4.19 will be done in a subsequent patch using
direct GIC_SPI interrupts.

Signed-off-by: Suman Anna <s-anna@ti.com>