arm64: dts: ti: k3-dra822-vlab: Add DDR carveout memory nodes for MCU R5Fs
Two carveout reserved memory nodes each have been added for each of the
MCU R5F remote processor devices within MCU domain on the K3 DRA822 VLAB
board. These nodes are assigned to the respective rproc device nodes as
well. The first region will be used as the DMA pool for the rproc devices,
and the second region will furnish the static carveout regions for the
firmware memory.

The current carveout addresses and sizes are defined statically for each
device. The R5F processors do not have an MMU, and as such require the
exact memory used by the firmwares to be set-aside. The firmware images
do not require any RSC_CARVEOUT entries in their resource tables to
allocate the memory for firmware memory segments.

Note that the R5F1 carveouts are needed only if the MCU is running in
Split (non-LockStep) mode. The reserved memory nodes can be disabled
later on if there is no use-case defined to use the corresponding
remote processor.

Signed-off-by: Suman Anna <s-anna@ti.com>