remoteproc/k3-r5: extend support to MCU R5 clusters on DRA82x SoCs
The K3 DRA82x SoCs typically have three MCU R5F clusters with 2
cores each. One cluster is present within the MCU voltage domain
(MCU0), and the remaining two clusters are present in the MAIN
voltage domain (MCU1 and MCU2).

Extend the support to these clusters in the K3 R5F remoteproc
driver using the DRA82x specific compatibles. The integration
of these clusters on DRA82x SoC is slightly different from those
on AM65x SoCs in that these IPs do support a local reset. The
current code sequence already supports these.

The names for the firmware images are fixed for each processor
and are expected to be as follows:
    MCU0 Core0: dra82x-mcu0-r5f0-fw (both in LockStep and Split modes)
    MCU0 Core1: dra82x-mcu0-r5f1-fw (needed only in Split mode)
    MCU1 Core0: dra82x-mcu1-r5f0-fw (both in LockStep and Split modes)
    MCU1 Core1: dra82x-mcu1-r5f1-fw (needed only in Split mode)
    MCU2 Core0: dra82x-mcu2-r5f0-fw (both in LockStep and Split modes)
    MCU2 Core1: dra82x-mcu0-r5f1-fw (needed only in Split mode)

Signed-off-by: Suman Anna <s-anna@ti.com>