TEMP: remoteproc/k3-r5: stub out unsupported ti-sci functions
The current kernel does not have the full support for the TI-SCI
protocol, and not all functions are available. Some of the required
DT properties are not yet available and have therefore been commented
out. Stub out the code parsing the unavailable DT properties using a
temporary macro that can easily be enabled if needed. This allows the
devices to be properly probed (with no auto-boot). All the ti-sci-proc
functions are also stubbed out to avoid any kernel panics if exercised.

Signed-off-by: Suman Anna <s-anna@ti.com>