drm/tidss: dispc7: Update mflag setup for AM654 and enable it
The mflag code was originally copied from dispc6 code, and while the
register accesses were changed to be correct on dispc7, the code only
handled a single plane. The mflag code was also intentionally never
called, as we had not verified the functionality.

This patch changes the code to handle multiple planes, sets the
MFLAG_CTRL to MFLAGEN mode instead of MFLAGFORCE (MFLAGFORCE was used
due to underflow issues on DRA7), and enables mflag support.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>