PCI: endpoint: Assign function number of each PF in EPC core
The endpoint core relies on the drivers that invoke the API to bind
a controller driver with a function driver to allocate and assign a
function number to each physical function. Ideally the function
numbers should be assinged by EPC core for each of the physical
function that is bound to the EPC.

Assign function number of each PF in EPC core and remove function
number allocation in endpoint configfs.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Tested-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>