arm64: dts: ti: dra822: Fix msi-map of PCIe RC dt node
The ITS device ID of the second PCI instance is also given as same
as the first PCI instance. This will cause issues if both instances
are used at the same time.
Also change the size of maximum size of device IDs to 0x10000 to allow
for all the supported bus numbers.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Tested-by: Jean-Jacques Hiblot <jjhiblot@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>