remoteproc/k3-r5: configure all R5s for IPC-only mode on DRA82x SoCs
Add support to the K3 R5F remoteproc driver to configure all the MCU
R5F cores on DRA82x SoCs to be in IPC-only mode. The IPC-only mode
expects that the remote processors are already booted by the bootloader,
and only perform the minimum steps required to initialize and deinitialize
the virtio IPC transports. Support for this is achieved through various
flags in both remoteproc core and the R5F remoteproc driver. The support
requires that the firmware still be requested to process the resource
table for retrieving the virtio device and trace resource entries.

NOTE:
- The current logic hard-codes the IPC-only mode for DRA82x SoCs, the
  ideal logic would have to rely on reading the status and state from
  the System Firmware and configure itself appropriately as either the
  remoteproc mode or IPC-only mode.
- The resource table address retrieval logic can actually be done by
  providing the data through device tree (thereby bypassing the need to
  request the firmware), with the bootloader adding and/or updating the
  necessary DT properties. This cannot be done for current kernel though
  as there is no bootloader and the early-boot is temporarily done through
  VLAB scripts.

Signed-off-by: Suman Anna <s-anna@ti.com>