ti_config_fragments: presil_baseport: Set number of UARTS to 16
The DRA82x platform uses the 8250_OMAP serial driver. The driver
is enabled through the standard ARMv8 defconfig file and inherits
the default in-kernel Kconfig value of 4 for the number of UARTS
it supports, but this is not enough to probe all the UARTS present
on the DRA822 platform. Set this number to 16 explicitly in the
presil_baseport.cfg file to ensure all the UARTs are probed properly.

Signed-off-by: Suman Anna <s-anna@ti.com>