remoteproc/k3-r5: add a remoteproc driver for MCU (dual-R5F) subsystem
The AM65x SoCs have a single Microcontroller Unit (MCU) subsystem that is
comprised of dual-core Arm Cortex-R5F processor cores. This MCU R5 cluster
is present within the MCU domain, and can be configured at boot time to be
either run in a LockStep mode or in an Asymmetric Multi Processing (AMP)
fashion in Split-mode. This subsystem has 64 KB each Tightly-Coupled Memory
(TCM) internal memories for each core split between two banks - TCMA and
TCMB (further interleaved into two banks). The subsystem does not have an
MMU, but a Region Address Translater (RAT) module that is accessible only
from the R5s for providing translations between 32-bit CPU addresses into
larger system bus addresses.

Add a remoteproc driver to support this subsystem to be able to load and
boot the R5 cores primarily in LockStep mode. The code also includes the
base for Split mode. Error Recovery and Power Management features are not
currently supported. Loading support includes the internal TCMs and DDR.
RAT support is left for a future patch, and as such the reserved memory
carveout regions are all expected to be using memory regions within the
first 2 GB.

The R5F remote processors do not have an MMU, and so require fixed memory
carveout regions matching the firmware image addresses. Support for this
is provided by mandating multiple memory regions to be attached to the
remoteproc device. The first memory region will be used to serve as the
DMA pool for all dynamic allocations like the vrings and vring buffers.
The remaining memory regions are mapped into the kernel at device probe
time, and are used to provide address translations for firmware image
segments without the need for any RSC_CARVEOUT entries. Any firmware
image using memory outside of the supplied reserved memory carveout
regions will be errored out.

The R5F processors on TI K3 SoCs require a specific sequence for booting
and shutting down the processors. This sequence is also dependent on the
mode (LockStep or Split) the R5F cluster is configured for. The R5F cores
have a Memory Protection Unit (MPU) that has a default configuration that
does not allow the cores to run out of DDR out of reset. This is resolved
by using the TCMs for boot-strapping code that applies the appropriate
executable permissions on desired DDR memory. The loading into the TCMs
requires that the resets be released first with the cores in halted state.
The Power Sleep Controller (PSC) module on K3 SoCs requires that the cores
be in WFI/WFE states with no active bus transactions before the cores can
be put back into reset. Support for this is provided by using the newly
introduced .prepare() and .unprepare() ops in the remoteproc core. The
.prepare() ops is invoked before any loading, and the .unprepare() ops
is invoked after the remoteproc resource cleanup. The R5F core resets
are deasserted in .prepare() and asserted in .unprepare(), and the cores
themselves are started and halted in .start() and .stop() ops. This
ensures symmetric usage and allows the R5F cores state machine to be
maintained properly between using the sysfs 'state' variable, bind/unbind
and regular module load/unload flows.

The subsystem is represented as a single remoteproc in LockStep mode, and
as two remoteprocs in Split mode. The driver uses various TI-SCI interfaces
to talk to the System Controller (DMSC) for managing configuration, power
and reset management of these cores. IPC between the A53 cores and the R5
cores is supported through the virtio rpmsg stack using shared memory and
OMAP Mailboxes.

The driver expects the following firmware names to be used for these
cores:
    am65x-mcu-r5f0-fw (LockStep mode and for Core0 in Split mode)
    am65x-mcu-r5f1-fw (Core1 in Split mode)

Signed-off-by: Suman Anna <s-anna@ti.com>