clk: keystone: sci-clk: workaround firmware limitation of 256 dev clks
SCI clock driver probes the available clocks on the device via the
firmware, by looking for number of parents of clocks. On DRA82x, the
number of board0 clocks is way over 300 though, which is impossible
to access via the APIs as the clk_id is hard limited to u8.

The firmware limitation causes the clock probing to be effectively
pruned at 8 bits, and resulting in a soft deadlock while probing
the board0 clocks.

Fix the issue by bailing out if clk_id reaches 255.

Signed-off-by: Tero Kristo <t-kristo@ti.com>