remoteproc/k3-dsp: lower the trace level on couple of debug prints
There are couple of informational debug traces that print the
kernel mapping and device addresses for the parsed internal memory
regions and fixed carveout, that are currently using KERN_ERR level
and so are being printed by default. These are only needed during
debug in general, so lower the trace level to KERN_DEBUG for both
of the traces to limit the default output.

Signed-off-by: Suman Anna <s-anna@ti.com>