remoteproc/k3-dsp: add boot sanity check for IPC-only mode
The commit 2bc5fbfd3aa4 ("remoteproc/k3-dsp: configure all DSPs for
IPC-only mode on DRA82x SoCs") has hard-wired all the DSPs on DRA82x
SoCs for IPC-only mode without checking if the bootloader has actually
started a DSP. The driver still would go through the processing of the
firmware and configure the virtio devices - this gives the impression
that the DSP was configured properly for IPC-only mode.

This patch queries the System Firmware and uses the DSP on status as
a means to determine the running status of the DSP before configuring
it for IPC-only mode. The driver probe will fail if the DSPs were
not started by the bootloader.

The IPC-only informational traces in the .start() and .stop() functions
are also revised slightly.

NOTE:
The local resets on VLAB are currently not modeled, and this logic
will be revised once the logic for the same is provided and supported.

Signed-off-by: Suman Anna <s-anna@ti.com>