remoteproc/k3-dsp: configure DSPs for remoteproc mode dynamically
The DSPs on J721E SoCs currently support IPC-only mode and requires that
the cores are loaded and started by bootloaders. The commit a30acd19376f
("remoteproc/k3-dsp: configure all DSPs for IPC-only mode on J721E SoCs")
performs a sanity check for the same and returns an error in the driver
probe function if a bootloader has not started a DSP core.

Use the same detection logic to configure the device for remoteproc mode
in the driver if a DSP is not booted by bootloader. The remoteproc mode
allows the kernel remoteproc driver to do the regular load and boot and
other device management operations for a DSP. This allows the driver to
be dynamically configured for either mode based on the DSP power on
status at driver probe time.

Note that the driver cannot configure a DSP for remoteproc mode by any
means without rebooting the kernel if that DSP has been started by a
bootloader.

Signed-off-by: Suman Anna <s-anna@ti.com>