remoteproc/k3-dsp: configure all DSPs for IPC-only mode on J721E SoCs
Add support to the K3 DSP remoteproc driver to configure all the C66x
and C71x cores on J721E SoCs to be in IPC-only mode. The IPC-only mode
expects that the remote processors are already booted by the bootloader,
and only perform the minimum steps required to initialize and deinitialize
the virtio IPC transports. Support for this is achieved through various
flags in both remoteproc core and the DSP remoteproc driver. The support
requires that the firmware still be requested to process the resource
table for retrieving the virtio device and trace resource entries.

The IPC-only mode is sanity checked by querying the System Firmware for
the DSP on state/status and making sure that the DSP is indeed started
by by the bootloaders, otherwise the driver probe will fail.

NOTE:
 - The current logic allows the driver to be probed successfully only for
   IPC-only mode for J721E SoCs. The logic to dynamically configure the
   driver for either IPC-only mode or remoteproc mode will be done in a
   subsequent patch.
 - The resource table address retrieval logic can actually be done by
   providing the data through device tree (thereby bypassing the need to
   request the firmware), with the bootloader adding and/or updating the
   necessary DT properties.
 - The local resets on VLAB are currently not modeled, and this logic
   will be revised once the logic for the same is provided and supported.

Signed-off-by: Suman Anna <s-anna@ti.com>