arm64: dts: ti: k3-j721e-vlab: Add DDR carveout memory nodes for MAIN R5Fs
Two carveout reserved memory nodes each have been added for each of the
R5F remote processor devices within MAIN domain on the K3 J721E VLAB
board. These nodes are assigned to the respective rproc device nodes as
well. The first region will be used as the DMA pool for the rproc devices,
and the second region will furnish the static carveout regions for the
firmware memory.

The current carveout addresses and sizes are defined statically for each
device. The R5F processors do not have an MMU, and as such require the
exact memory used by the firmwares to be set-aside. The firmware images do
not require any RSC_CARVEOUT entries in their resource tables to allocate
the memory for firmware memory segments.

Note that the R5F1 carveouts are needed only if the R5F cluster is running
in Split (non-LockStep) mode. The reserved memory nodes can be disabled
later on if there is no use-case defined to use the corresponding remote
processor.

Signed-off-by: Suman Anna <s-anna@ti.com>