misc: pci_endpoint_test: Avoid using module parameter to determine irqtype
commit e03327122e2c8e6ae4565ef ("pci_endpoint_test: Add 2 ioctl
commands") in the pci_endpoint_test_set_irq() uses module parameter
'irqtype' to check if irq vectors of a particular type is already
allocated. However with multi-function devices, irqtype will not
correctly reflect the irq type of the PCI device. Fix it here by
adding 'irqtype' for each PCI device to shows the irq type of that
particular PCI device.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>