arm64: dts: ti: k3-j721e-vlab: Swap DMA reserved memory nodes between C66x DSPs
Each C66x DSP currently uses two carveout reserved memory nodes, with
the first region used as the DMA pool for the rproc devices (used for
allocating vrings, vring buffers etc) and the second region used for
the static carveout region for the firmware memory. These two regions
are next to each other and together make up 16 MB of size.

The minimum granularity on the Cache settings on C66x DSP cores is 16 MB,
so it is not possible to set different cache attributes to the above
regions. Resolve this issue by simply swapping the DMA memory regions
between the two C66x DSP cores - this ensures the DMA region and firmware
memory regions are in separate 16 MB regions, and not affecting the overall
remoteproc carveout regions.

Note that this breaks backward compatibility and requires new DSP firmwares
with updated IPC regions on the RTOS side.

Signed-off-by: Suman Anna <s-anna@ti.com>