TEMP: remoteproc/k3-dsp: Deny auto-boot in remoteproc mode
The K3 DSP remoteproc driver was designed to support IPC-only mode
originally, and the driver errored out if the DSPs were not booted by
U-boot. The boot detection logic was reused to configure the driver
for remoteproc mode in commit 72b021c49a08 ("remoteproc/k3-dsp:
configure DSPs for remoteproc mode dynamically").

This causes increased boot-times on J721E VLAB platform when the DSPs
are booted automatically during module loading and slow down any further
tests/applications. So, configure the DSPs to not auto-boot specifically
in remoteproc mode to mitigate this slowdown, and without reverting the
functionality. This allows the remoteprocs to be started and stopped
only on a demand basis through remoteproc sysfs 'state' file.

Signed-off-by: Suman Anna <s-anna@ti.com>