drm/tidss: HACK: Put dispc_ops->runtime_resume() within ifndef CONFIG_PM
The explicit calls to dispc runtime functions within #ifndef CONFIG_PM
is for a case when we are running in an environment that does not
support power management and the IP is always powered.  This is the
case with some simulation environments. In such a case we still need
to initialize dispc registers, before starting the normal operation.

Before this commit those calls were enclosed within #ifndef
CONFIG_PM_SLEEP, which was broken since pm_runtime support is
configured with CONFIG_PM.

Signed-off-by: Jyri Sarha <jsarha@ti.com>