arm64: dts: ti: k3-j721e-main: Fix "bus-range" and "iommu-map" of PCIe RC dt node
Now that we use non BDF mode, shift back to using bus-range starting
from '0'. Also fix iommu-map to provide the correct streamID to SMMU.
The upper '4' bits of stream ID is fixed for each PCIe instance. For
the 1st instance the upper 4 bits of streamID is '0b0000' and for the
second instance the upper 4 bits of streamID is '0b0100'

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>